discriminator propertyType
